+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 0.11.11
 * Optimize
   - register `userAreaSelection` will be warn when `props.useAreaSelection` not provide
+* Bugfix
+  - fix: `highlightDimCell` can not remove className when in css method.
 
 ## 0.11.10
 * Feature